+++ src/Sentry/Internal/Extensions/JsonExtensions.cs
@@ -49,6 +49,26 @@ internal static class JsonExtensions
         return result;
     }
 
+    public static Dictionary<string, TValue>? GetDictionaryOrNull<TValue>(
+        this JsonElement json,
+        Func<JsonElement, TValue> factory)
+        where TValue : IJsonSerializable?
+    {
+        if (json.ValueKind != JsonValueKind.Object)
+        {
+            return null;
+        }
+
+        var result = new Dictionary<string, TValue>();
+
+        foreach (var (name, value) in json.EnumerateObject())
+        {
+            result[name] = factory(value);
+        }
+
+        return result;
+    }
+
     public static Dictionary<string, string?>? GetStringDictionaryOrNull(this JsonElement json)
     {
         if (json.ValueKind != JsonValueKind.Object)
@@ -136,7 +156,7 @@ internal static class JsonExtensions
 #if NETCOREAPP || NETSTANDARD2_1_OR_GREATER
         var substring = s[2..];
 #else
-            var substring = s.Substring(2);
+        var substring = s.Substring(2);
 #endif
         if (s.StartsWith("0x") &&
             long.TryParse(substring, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out var result))
@@ -186,6 +206,37 @@ internal static class JsonExtensions
         }
     }
 
+    public static void WriteDictionaryValue<TValue>(
+        this Utf8JsonWriter writer,
+        IEnumerable<KeyValuePair<string, TValue>>? dic,
+        IDiagnosticLogger? logger,
+        bool includeNullValues = true)
+        where TValue : IJsonSerializable?
+    {
+        if (dic is not null)
+        {
+            writer.WriteStartObject();
+
+            foreach (var (key, value) in dic)
+            {
+                if (value is not null)
+                {
+                    writer.WriteSerializable(key, value, logger);
+                }
+                else if (includeNullValues)
+                {
+                    writer.WriteNull(key);
+                }
+            }
+
+            writer.WriteEndObject();
+        }
+        else
+        {
+            writer.WriteNullValue();
+        }
+    }
+
     public static void WriteStringDictionaryValue(
         this Utf8JsonWriter writer,
         IEnumerable<KeyValuePair<string, string?>>? dic)
@@ -212,6 +263,17 @@ internal static class JsonExtensions
         string propertyName,
         IEnumerable<KeyValuePair<string, object?>>? dic,
         IDiagnosticLogger? logger)
+    {
+        writer.WritePropertyName(propertyName);
+        writer.WriteDictionaryValue(dic, logger);
+    }
+
+    public static void WriteDictionary<TValue>(
+        this Utf8JsonWriter writer,
+        string propertyName,
+        IEnumerable<KeyValuePair<string, TValue>>? dic,
+        IDiagnosticLogger? logger)
+        where TValue : IJsonSerializable?
     {
         writer.WritePropertyName(propertyName);
         writer.WriteDictionaryValue(dic, logger);
@@ -549,10 +611,24 @@ internal static class JsonExtensions
         IEnumerable<KeyValuePair<string, object?>>? dic,
         IDiagnosticLogger? logger)
     {
-        var asDictionary = dic as IReadOnlyDictionary<string, object?> ?? dic?.ToDictionary();
-        if (asDictionary is not null && asDictionary.Count > 0)
+        var dictionary = dic as IReadOnlyDictionary<string, object?> ?? dic?.ToDictionary();
+        if (dictionary is not null && dictionary.Count > 0)
         {
-            writer.WriteDictionary(propertyName, asDictionary, logger);
+            writer.WriteDictionary(propertyName, dictionary, logger);
+        }
+    }
+
+    public static void WriteDictionaryIfNotEmpty<TValue>(
+        this Utf8JsonWriter writer,
+        string propertyName,
+        IEnumerable<KeyValuePair<string, TValue>>? dic,
+        IDiagnosticLogger? logger)
+        where TValue : IJsonSerializable?
+    {
+        var dictionary = dic as IReadOnlyDictionary<string, TValue> ?? dic?.ToDictionary();
+        if (dictionary is not null && dictionary.Count > 0)
+        {
+            writer.WriteDictionary(propertyName, dictionary, logger);
         }
     }
 
@@ -561,10 +637,10 @@ internal static class JsonExtensions
         string propertyName,
         IEnumerable<KeyValuePair<string, string?>>? dic)
     {
-        var asDictionary = dic as IReadOnlyDictionary<string, string?> ?? dic?.ToDictionary();
-        if (asDictionary is not null && asDictionary.Count > 0)
+        var dictionary = dic as IReadOnlyDictionary<string, string?> ?? dic?.ToDictionary();
+        if (dictionary is not null && dictionary.Count > 0)
         {
-            writer.WriteStringDictionary(propertyName, asDictionary);
+            writer.WriteStringDictionary(propertyName, dictionary);
         }
     }
 
@@ -574,10 +650,10 @@ internal static class JsonExtensions
         IEnumerable<object?>? arr,
         IDiagnosticLogger? logger)
     {
-        var asList = arr as IReadOnlyList<object?> ?? arr?.ToArray();
-        if (asList is not null && asList.Count > 0)
+        var list = arr as IReadOnlyList<object?> ?? arr?.ToArray();
+        if (list is not null && list.Count > 0)
         {
-            writer.WriteArray(propertyName, asList, logger);
+            writer.WriteArray(propertyName, list, logger);
         }
     }
 
@@ -586,10 +662,10 @@ internal static class JsonExtensions
         string propertyName,
         IEnumerable<string?>? arr)
     {
-        var asList = arr as IReadOnlyList<string?> ?? arr?.ToArray();
-        if (asList is not null && asList.Count > 0)
+        var list = arr as IReadOnlyList<string?> ?? arr?.ToArray();
+        if (list is not null && list.Count > 0)
         {
-            writer.WriteStringArray(propertyName, asList);
+            writer.WriteStringArray(propertyName, list);
         }
     }