+++ src/Sentry/TransactionTracer.cs
@@ -1,12 +1,14 @@
 using System.Collections.Concurrent;
+using System.ComponentModel;
 using Sentry.Internal;
+using Sentry.Protocol;
 
 namespace Sentry;
 
 /// <summary>
 /// Transaction tracer.
 /// </summary>
-public class TransactionTracer : ITransaction, IHasDistribution, IHasTransactionNameSource
+public class TransactionTracer : ITransaction, IHasDistribution, IHasTransactionNameSource, IHasMeasurements
 {
     private readonly IHub _hub;
     private readonly SentryStopwatch _stopwatch = SentryStopwatch.StartNew();
@@ -162,6 +164,11 @@ public class TransactionTracer : ITransaction, IHasDistribution, IHasTransaction
     /// <inheritdoc />
     public IReadOnlyCollection<ISpan> Spans => _spans;
 
+    private readonly ConcurrentDictionary<string, Measurement> _measurements = new();
+
+    /// <inheritdoc />
+    public IReadOnlyDictionary<string, Measurement> Measurements => _measurements;
+
     /// <inheritdoc />
     public bool IsFinished => EndTimestamp is not null;
 
@@ -220,6 +227,11 @@ public class TransactionTracer : ITransaction, IHasDistribution, IHasTransaction
     /// <inheritdoc />
     public void UnsetTag(string key) =>
         _tags.TryRemove(key, out _);
+
+    /// <inheritdoc />
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    public void SetMeasurement(string name, Measurement measurement) =>
+        _measurements[name] = measurement;
 
     internal ISpan StartChild(SpanId parentSpanId, string operation)
     {